fix broken filtering
